Update 85872959.geojson
Removing 'District' from label for City of Aurora, which is not a term used to describe this area. As this data flows into systems such as Geocode.Earth, which is used by our School Choice Vendor, an address within Aurora shows a result of 'Aurora District', which causes additional confusion in that it suggests the Aurora School District, which is not the correct name either.